Moved Code To relevant Places
