+++ ProjectGallery/ClashOfClansHelper/Models/SingleBuilding.cs
