+++ ProjectGallery/ClashOfClansHelper/Models/SingleBuildingImage.cs
