+++ ProjectGallery/ClashOfClansHelper/Models/SingleBuildingSingleLine.cs
