+++ ProjectGallery/ClashOfClansHelper/Utilites/ImageLoader.cs
