+++ ProjectGallery/LinqDemoData/MainWindow.xaml.cs
@@ -11,6 +11,7 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
+using LinqDemoData.Models;
 
 namespace LinqDemoData;
