+++ ProjectGallery/LinqDemoData/Models/Products.cs
@@ -1,4 +1,4 @@
-﻿namespace LinqDemoData;
+﻿namespace LinqDemoData.Models;
 
 class Products
 {
