+++ ProjectGallery/UsersAPI/Models/Users.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 
-namespace UsersAPI
+namespace UsersAPI.Models
 {
     public class UsersResponse
     {
