+++ ProjectGallery/UsersCRUDApi/MainWindow.xaml.cs
@@ -10,6 +10,7 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
+using UsersCRUDApi.Models;
 using static System.Net.WebRequestMethods;
 
 namespace UsersCRUDApi
