+++ ProjectGallery/UsersCRUDApi/Models/Users.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 
-namespace UsersCRUDApi;
+namespace UsersCRUDApi.Models;
 
 public class UsersResponse
 {
